backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tokenizers = "^0.19"
transformers = "^4.42.4"
accelerate = "1.1.1"
aiohappyeyeballs = "2.4.3"
aiohttp = "3.11.8"
aiosignal = "1.3.1"
annotated-types = "0.7.0"
anyio = "4.6.2.post1"
attrs = "24.2.0"
audioread = "3.0.1"
av = "13.1.0"
build = "1.2.2.post1"
cachecontrol = "0.14.1"
chattts = "0.2.1"
cleo = "2.1.0"
click = "8.1.7"
crashtest = "0.4.1"
cryptography = "44.0.0"
datasets = "3.1.0"
decorator = "5.1.1"
dill = "0.3.8"
distlib = "0.3.9"
dulwich = "0.21.7"
einops = "0.8.0"
einx = "0.3.0"
encodec = "0.1.1"
fastapi = "0.115.5"
fastjsonschema = "2.21.1"
frozendict = "2.4.6"
frozenlist = "1.5.0"
fsspec = "2024.9.0"
h11 = "0.14.0"
httpcore = "1.0.7"
httpx = "0.28.0"
huggingface-hub = "0.26.3"
installer = "0.7.0"
jaraco-classes = "3.4.0"
jeepney = "0.8.0"
joblib = "1.4.2"
keyring = "24.3.1"
lazy-loader = "0.4"
librosa = "0.10.2.post1"
llvmlite = "0.43.0"
lxml = "5.3.0"
more-itertools = "10.5.0"
msgpack = "1.1.0"
multidict = "6.1.0"
multiprocess = "0.70.16"
nvidia-cublas-cu12 = "12.4.5.8"
nvidia-cuda-cupti-cu12 = "12.4.127"
nvidia-cuda-nvrtc-cu12 = "12.4.127"
nvidia-cuda-runtime-cu12 = "12.4.127"
nvidia-cudnn-cu12 = "9.1.0.70"
nvidia-cufft-cu12 = "11.2.1.3"
nvidia-curand-cu12 = "10.3.5.147"
nvidia-cusolver-cu12 = "11.6.1.9"
nvidia-cusparse-cu12 = "12.3.1.170"
nvidia-nvjitlink-cu12 = "12.4.127"
pandas = "2.2.3"
pexpect = "4.9.0"
pkginfo = "1.12.0"
platformdirs = "4.3.6"
poetry = "1.8.5"
poetry-core = "1.9.1"
poetry-plugin-export = "1.8.0"
pooch = "1.8.2"
propcache = "0.2.0"
psutil = "6.1.0"
ptyprocess = "0.7.0"
pyarrow = "18.1.0"
pybase16384 = "0.3.7"
pydantic = "2.10.2"
pydantic-core = "2.27.1"
pydub = "0.25.1"
pyproject-hooks = "1.2.0"
python-dateutil = "2.9.0.post0"
python-multipart = "0.0.18"
pytils = "0.4.1"
pytz = "2024.2"
rapidfuzz = "3.10.1"
regex = "2024.11.6"
requests-toolbelt = "1.0.0"
scikit-learn = "1.5.2"
scipy = "1.14.1"
secretstorage = "3.3.3"
setuptools = "75.1.0"
shellingham = "1.5.4"
six = "1.16.0"
sniffio = "1.3.1"
soundfile = "0.12.1"
soxr = "0.5.0.post1"
starlette = "0.41.3"
sympy = "1.13.1"
threadpoolctl = "3.5.0"
tomlkit = "0.13.2"
trove-classifiers = "2024.10.21.16"
tzdata = "2024.2"
uvicorn = "0.32.1"
vector-quantize-pytorch = "1.20.10"
virtualenv = "20.28.0"
vocos = "0.1.0"
wheel = "0.44.0"
xxhash = "3.5.0"
yarl = "1.18.0"
pillow = "^11.0.0"
facexlib = "^0.3.0"
sentencepiece = "^0.2.0"
diffusers = "^0.31.0"
realesrgan = "^0.3.0"
torchvision = "0.20.1"
torch = "2.5.1"
ffmpeg-python = "^0.2.0"
pysrt = "^1.1.2"
openai-whisper = "^20240930"
srt = "^3.5.3"
jieba = "^0.42.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

